Move get_attr_value() method to DeadlineDefaultJobAttrs() class
